requirements3.txt
dj-database-url==0.4.1
Django==1.9.7
django-environ==0.4.1
telepot==10.5
gunicorn==19.6.0
psycopg2==2.6.1
whitenoise==2.0.6
requests==2.9.1
